backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
sqlparse==0.5.1
typing_extensions==4.12.2
djangorestframework-simplejwt==5.3.1
qrcode==8.0
pyotp==2.9.0
pillow==10.4.0
